Separate database generation and visualization.
I had originally hoped that the MAP-Client would be able to keep a list of all know MAP-Client plugins and would be able to check the validity of this list by calling the GitHub API directly. Unfortunately, due to GitHub API rate limiting, it's not possible to retrieve the relevant MAP-Client plugin information directly from the GitHub API unless the user is authenticated. The generation of the plugin database is now done separately from the visualization of this data.
